backend/requirements.txt
gradio==4.44.0
openai==1.54.0
groq==0.11.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
aiohttp==3.10.5
